requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
kafka-python==2.0.1
pymsteams==0.1.13
requests==2.24.0
urllib3==1.25.10
